+++ packages/blocks/core/js/wordpress/paragraph/index.js
@@ -15,7 +15,13 @@ export const Paragraph: BlockType = {
 	name: 'blockeraParagraph',
 	targetBlock: 'core/paragraph',
 	blockeraInnerBlocks: {
-		'elements/link': sharedInnerBlocks['elements/link'],
+		'elements/link': {
+			...sharedInnerBlocks['elements/link'],
+			settings: {
+				...sharedInnerBlocks['elements/link'].settings,
+				force: false,
+			},
+		},
 	},
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;