+++ packages/blocks/core/CHANGELOG.md
@@ -32,6 +32,7 @@
 - `Home Link` block: inner blocks improved.
 - `Navigation Link` block: inner blocks improved.
 - `Submenu` block: inner blocks improved.
+- `Paragraph` block: inner blocks improved.
 
 
 ### Automated Tests